\name{armadp-package}
\alias{armadp-package}
\alias{armadp}
\docType{package}
\title{\packageTitle{armadp}}
\description{\packageDescription{armadp}}
\section{Package Content}{\packageIndices{armadp}}
\author{\packageAuthor{armadp}}
\section{Maintainer}{\packageMaintainer{armadp}}
\keyword{package}
